+++ app/src/main/java/com/example/hw1/MainFragment.java
@@ -47,6 +47,7 @@ public class MainFragment extends Fragment implements Constants, DialogResult {
                              Bundle savedInstanceState) {
         return inflater.inflate(R.layout.fragment_main, container, false);
 
+
     }
